+++ src/js/app/app.js
@@ -1,5 +1,0 @@
-define(function(require) {
-  var Application = require('lavaca/mvc/Application');
-  var app = new Application(function() {});
-  return app;
-});