+++ src/mvc/Application.js
@@ -10,7 +10,6 @@ define(function(require) {
     Template = require('lavaca/ui/Template'),
     Config = require('lavaca/util/Config'),
     Promise = require('lavaca/util/Promise'),
-    ChildBrowser = require('lavaca/env/ChildBrowser'),
     Translation = require('lavaca/util/Translation');
 
   function _stopEvent(e) {
@@ -24,6 +23,21 @@ define(function(require) {
       return matches[1].replace(/#.*/, '');
     }
     return null;
+  }
+
+  function _isExternal(url) {
+    var match = url.match(/^([^:\/?#]+:)?(?:\/\/([^\/?#]*))?([^?#]+)?(\?[^#]*)?(#.*)?/);
+    if (typeof match[1] === 'string' 
+        && match[1].length > 0 
+        && match[1].toLowerCase() !== location.protocol) {
+      return true;
+    }
+    if (typeof match[2] === 'string'
+        && match[2].length > 0
+        && match[2].replace(new RegExp(':('+{'http:':80,'https:':443}[location.protocol]+')?$'), '') !== location.host) {
+      return true;
+    }
+    return false;
   }
 
   /**
@@ -101,28 +115,37 @@ define(function(require) {
      * @param {Event} e  The event object
      */
     onTapLink: function(e) {
-      var link = $(e.currentTarget),
-          url = link.attr('href'),
-          rel = link.attr('rel'),
-          target = link.attr('target'),
-          isExternal = link.is('[data-external]');
-      if (/^((mailto)|(tel)|(sms))\:/.test(url)) {
-        location.href = url;
-        return true;
-      } else {
-        e.preventDefault();
-      }
-      if (rel === 'back') {
-        History.back();
-      } else if (isExternal || rel === 'nofollow' || target === '_blank') {
-        e.stopPropagation();
-        new ChildBrowser().showWebPage(url);
-      } else if (rel === 'cancel') {
-        this.viewManager.dismiss(e.currentTarget);
-      } else if (url) {
-        this.router.exec(url, null, null).error(this.onInvalidRoute);
-      }
-    },
+      var link = $(e.currentTarget),
+          defaultPrevented = e.isDefaultPrevented(),
+          url = link.attr('href') || link.attr('data-href'),
+          rel = link.attr('rel'),
+          target = link.attr('target'),
+          isExternal = link.is('[data-external]') || _isExternal(url);
+
+      if (!defaultPrevented) {
+        if (Device.isCordova() && target) {
+          e.preventDefault();
+          window.open(url, target || '_blank');
+        } else if (isExternal || target || (e.ctrlKey || e.metaKey)) {
+          return true;
+        } else {
+          e.preventDefault();
+          if (rel === 'back') {
+            History.back();
+          } else if (rel === 'force-back' && url) {
+            History.isRoutingBack = true;
+            this.router.exec(url, null, null).always(function() {
+              History.isRoutingBack = false;
+            });
+          } else if (rel === 'cancel') {
+            this.viewManager.dismiss(e.currentTarget);
+          } else if (url) {
+            url = url.replace(/^\/?#/, '');
+            this.router.exec(url).error(this.onInvalidRoute);
+          }
+        }
+      }
+    },
     /**
      * Makes an AJAX request if the user is online. If the user is offline, the returned
      * promise will be rejected with the string argument "offline". (Alias for [[Lavaca.net.Connectivity]].ajax)
@@ -196,13 +219,15 @@ define(function(require) {
      * @method bindLinkHandler
      */
     bindLinkHandler: function() {
-      var $body = $(document.body);
+      var $body = $(document.body),
+          type = 'click';
       if ($body.hammer) {
         $body = $body.hammer();
+        type = 'tap';
       }
       $body
-        .on('tap click', '.ui-blocker', _stopEvent)
-        .on('tap click', 'a', this.onTapLink.bind(this));
+        .on(type, '.ui-blocker', _stopEvent)
+        .on(type, 'a', this.onTapLink.bind(this));
     },
     /**
      * Gets initial route based on query string returned by server 302 redirect